+++ cogs/shared/handler.py
@@ -1,5 +1,5 @@
-'''
-Discord Raidkit v2.3.0 — "The trojan horse of discord raiding"
+"""
+Discord Raidkit v2.3.1 — "The trojan horse of discord raiding"
 Copyright © 2022 the-cult-of-integral
 
 a collection of raiding tools, hacking tools, and a token grabber generator for discord; written in Python 3
@@ -8,11 +8,12 @@ This program is under the GNU General Public License v2.0.
 https://github.com/the-cult-of-integral/discord-raidkit/blob/master/LICENSE
 
 handler.py contains the on_command_error cog listener.
-handler.py was last updated on 11/08/22 at 17:53.
-'''
+handler.py was last updated on 04/09/22 at 18:51.
+"""
 
 import discord
 from discord.ext import commands
+
 
 class Handler(commands.Cog):
     def __init__(self, bot: commands.Bot) -> None:
@@ -57,6 +58,7 @@ class Handler(commands.Cog):
                 description='**An unknown error has occurred.**',
                 color=discord.Color.brand_red())
             await ctx.send(embed=embed)
+
 
 async def setup(bot: commands.Bot) -> None:
     await bot.add_cog(Handler(bot))